Files reorganization: .hpp added (for template classes)
