--- IlluminatiConfirmed/task2/Array.cpp
@@ -1,8 +1,0 @@
-#include "Array.h"
-
-using IlluminatiConfirmed::Array;
-
-Array::Array() : m_data(NULL), m_size(0), m_capacity(0)
-{
-
-}
