+++ IlluminatiConfirmed/task2/task2.pro
@@ -4,9 +4,8 @@ CONFIG -= app_bundle
 CONFIG -= qt
 
 SOURCES += main.cpp \
-    #Array.cpp \
-    Array.t.cpp \
-    Vector.t.cpp
+    Array/Array.t.cpp \
+    Vector/Vector.t.cpp
     
 # Google Test
 unix|win32: LIBS += -L$$PWD/../../../googletest-master/googletest/lib/ -lgtest
@@ -14,10 +13,18 @@ unix|win32: LIBS += -L$$PWD/../../../googletest-master/googletest/lib/ -lgtest
 INCLUDEPATH += $$PWD/../../../googletest-master/googletest/include
 DEPENDPATH += $$PWD/../../../googletest-master/googletest/include
 
+INCLUDEPATH += Vector \
+            Array \
+            ContainerInterface
+
 HEADERS += \
-    Array.h \
-    Vector.h \
-    ContainerInterface.h \
-    Iterator.h \
-    BugNumbers.h
+    BugNumbers.h \
+    ContainerInterface.hpp \
+    Array/Array.h \
+    Array/Array.hpp \
+    ContainerInterface/ContainerInterface.h \
+    ContainerInterface/ContainerInterface.hpp \
+    Vector/Vector.h \
+    Vector/Vector.hpp \
+    Iterator.h
